Fix: DataArrayItem for stream to reg should be consistently uint64_t, for now

+++ source/hal/inc/cheby_gen/stream_to_reg_cpp.h
@@ -73,7 +73,7 @@ namespace myModule
             using MemReg::MemReg;
         };
 
-        struct DataArrayItem : MemReg<StreamToReg, AccessMode::RO, double>
+        struct DataArrayItem : MemReg<StreamToReg, AccessMode::RO, uint64_t>
         {
             using MemReg::MemReg;
         };
